Complete Laravel hybrid interview test implementation
- Implemented MerchantService with register, update, find, and payout methods
- Implemented AffiliateService with register method and email validation
- Implemented OrderService with order processing and affiliate registration
- Implemented PayoutOrderJob with transaction handling
- Implemented WebhookController and MerchantController
- Fixed database migrations to use decimal instead of float for monetary values
- Added required fields to Order model
- All tests passing (15/15) while respecting no modification of tests/factories rule

+++ database/migrations/2022_05_13_220658_create_affiliates_table.php
@@ -17,8 +17,8 @@ return new class extends Migration
             $table->id();
             $table->foreignId('user_id');
             $table->foreignId('merchant_id');
-            // TODO: Replace me with a brief explanation of why floats aren't the correct data type, and replace with the correct data type.
-            $table->float('commission_rate');
+            // Floats are not suitable for monetary values due to precision issues. Use decimal instead.
+            $table->decimal('commission_rate', 5, 4);
             $table->string('discount_code');
             $table->timestamps();
         });
